+++ docs/CHANGELOG.md
@@ -24,6 +24,7 @@ Changes are grouped as follows:
 - Verification of reversed connections was done of `property (linage)` instead of `view_property`
 ### Changed
 - [BREAKING] Rules are simplified and normalized, especially metadata
+- [BREAKING] removed `reference`, `class (linage)` and `property (linage)`.
 
 
 ## [0.97.3] - 16-11-**2024**